+++ cypress/e2e/a11y-en.cy.js
@@ -1,10 +1,8 @@
 /// <reference types="cypress" />
 
 const enLinks = require('../../src/en/en.json');
-const frLinks = require('../../src/fr/fr.json');
 
 const pagesEn = [];
-const pagesFr = [];
 
 // Create index of English pages
 Object.keys(enLinks.links).forEach(key => {
@@ -39,39 +37,6 @@ Object.keys(enLinks.links).forEach(key => {
   }
 });
 
-// Create index of French pages
-Object.keys(frLinks.links).forEach(key => {
-  const url = frLinks.links[key];
-  if (
-    !url.includes('developpement-en-cours') &&
-    !url.includes('https') &&
-    !url.includes('mailto') &&
-    !url.includes('demo')
-  ) {
-    let regex = /composants\/[a-z]/;
-    const pageName = key.replace(/([A-Z])/g, ' $1');
-    if (regex.test(url)) {
-      pagesFr.push({
-        name: `FR - ${pageName} - use case`,
-        url,
-      });
-      pagesFr.push({
-        name: `FR - ${pageName} - design`,
-        url: `${url}/design/`,
-      });
-      pagesFr.push({
-        name: `FR - ${pageName} - code`,
-        url: `${url}/code/`,
-      });
-    } else {
-      pagesFr.push({
-        name: `FR - ${pageName}`,
-        url,
-      });
-    }
-  }
-});
-
 describe(`A11Y test English documentation site`, () => {
   for (const page of pagesEn) {
     it(`${page.name}: ${page.url}`, () => {
@@ -81,23 +46,6 @@ describe(`A11Y test English documentation site`, () => {
         cy.checkA11y(null, null, terminalLog);
         // skip theme and topic menu since links are pulled from external source
         if (!page.url.includes('theme-and-topic-menu')) {
-          cy.scanDeadLinks();
-        }
-      });
-    });
-  }
-});
-
-describe(`A11Y test French documentation site`, () => {
-  after
-  for (const page of pagesFr) {
-    it(`${page.name}: ${page.url}`, () => {
-      cy.visit(page.url);
-      cy.get('.hydrated').then(() => {
-        cy.injectAxe();
-        cy.checkA11y(null, null, terminalLog);
-        // skip theme and topic menu since links are pulled from external source
-        if (!page.url.includes('menu-thematique')) {
           cy.scanDeadLinks();
         }
       });
